--- a/python/taichi/lang/util.py
+++ b/python/taichi/lang/util.py
@@ -11,25 +11,6 @@ from taichi.lang import impl
 from taichi.types.primitive_types import (f16, f32, f64, i8, i16, i32, i64, u8,
                                           u16, u32, u64)
 
-_has_pytorch = False
-_has_paddle = False
-
-_env_torch = os.environ.get('TI_ENABLE_TORCH', '1')
-if not _env_torch or int(_env_torch):
-    try:
-        import torch
-        _has_pytorch = True
-    except:
-        pass
-
-_env_paddle = os.environ.get('TI_ENABLE_PADDLE', '1')
-if not _env_paddle or int(_env_paddle):
-    try:
-        import paddle
-        _has_paddle = True
-    except:
-        pass
-
 
 def has_pytorch():
     """Whether has pytorch in the current Python environment.
@@ -38,6 +19,14 @@ def has_pytorch():
         bool: True if has pytorch else False.
 
     """
+    _has_pytorch = False
+    _env_torch = os.environ.get('TI_ENABLE_TORCH', '1')
+    if not _env_torch or int(_env_torch):
+        try:
+            import torch  # pylint: disable=C0415
+            _has_pytorch = True
+        except:
+            pass
     return _has_pytorch
 
 
@@ -47,26 +36,37 @@ def has_paddle():
     Returns:
         bool: True if has paddle else False.
     """
+    _has_paddle = False
+    _env_paddle = os.environ.get('TI_ENABLE_PADDLE', '1')
+    if not _env_paddle or int(_env_paddle):
+        try:
+            import paddle  # pylint: disable=C0415
+            _has_paddle = True
+        except:
+            pass
     return _has_paddle
 
 
-from distutils.spawn import find_executable
-
-# Taichi itself uses llvm-10.0.0 to compile.
-# There will be some issues compiling CUDA with other clang++ version.
-_clangpp_candidates = ['clang++-10']
-_clangpp_presence = None
-for c in _clangpp_candidates:
-    if find_executable(c) is not None:
-        _clangpp_presence = find_executable(c)
-
-
-def has_clangpp():
-    return _clangpp_presence is not None
+_clangpp_presence = False
 
 
 def get_clangpp():
-    return _clangpp_presence
+    from distutils.spawn import find_executable  # pylint: disable=C0415
+
+    # Taichi itself uses llvm-10.0.0 to compile.
+    # There will be some issues compiling CUDA with other clang++ version.
+    _clangpp_candidates = ['clang++-10', 'clang++']
+    for c in _clangpp_candidates:
+        if find_executable(c) is not None:
+            _clangpp_presence = find_executable(c)
+            return _clangpp_presence
+    return None
+
+
+def has_clangpp():
+    if _clangpp_presence is False:
+        get_clangpp()
+    return _clangpp_presence is not None
 
 
 def is_matrix_class(rhs):