Fix #7711 

Reduce import overhead by:
* Don't import pyTorch when importing Taichi. (It will be imported when necessary)
* Don't unconditionally get clangpp executable path.

The former would cutdown Taichi import overhead from ~2s down to 0.7s, while the latter further reduce it from 0.7s to 0.5s.